Always show in-app language picker (regardless of Android versions)
Hans raises some very good questions during review feedback, namely on how it presents or means to an ordinary user.
IMHO the thing with Android's per app language picker, or anything other than our own implementation for that matter (so it's not entirely its fault), is that inherently, we cannot be sure it presents all our available locales since it's simply 'out of our hands'.
At second thought, when we present a list to the users, we probably want to present them a full list (rather than some arbitrarily or inadvertently curated list) so that they can make an unfettered and informed choice, by default; and the only way to (reliably) do so seems to be to roll our own.
That said; code-wise the two 'plays nicely' with each other (since both proxy to essentially the same system APIs on Android 13+) so I'd venture to think it's a UX issue.  As much as we'd like users to enjoy a seamless Android experience, it's just that we do not need to 'encourage' or 'discourage' a particular Android feature on this occasion: users would be 'smart enough' to figure out per app language preferences in system settings on their own, and if they don't, they probably don't have to since our in-app language picker does a pretty decent job (of course).